Move grad and jacobian
